radiomics_project/requirements.txt
# 项目依赖

# 基础库
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# 医学影像处理
SimpleITK>=2.3.0
pydicom>=2.4.0
nibabel>=5.1.0
pyradiomics>=3.0.0

# 机器学习
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
imbalanced-learn>=0.11.0

# 深度学习
torch>=2.0.0
torchvision>=0.15.0
monai>=1.3.0

# 数据下载
requests>=2.31.0

# 可解释性
shap>=0.42.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0

# 代码质量
black>=23.0.0
flake8>=6.1.0
